Some refactoring to fix build problems
