+++ app/[lang]/page.tsx
@@ -1,14 +1,44 @@
+'use client';
 import { Box } from '@mui/material';
-import Dashboard from '@/app/[lang]/dashboard/page';
+import IntroScreen from '@/app/components/IntroScreen';
 
-const Home = async ({ params: { lang } }: { params: { lang: string } }) => {
+import { useEffect, useState } from 'react';
+import { getCookie } from 'cookies-next';
+import { getDictionary } from '@/app/[lang]/dictionaries';
+import { useRouter } from 'next/navigation';
+
+const Home = ({ params: { lang } }: { params: { lang: string } }) => {
+  const [dict, setDict] = useState<Dict>({});
+  const [isMounted, setIsMounted] = useState(false);
+  const router = useRouter();
+
+  useEffect(() => {
+    setIsMounted(true);
+    if (lang) {
+      getDictionary(lang as Lang).then((res) => {
+        if (res) {
+          setDict(res);
+        }
+      });
+    }
+    if (getCookie('language')) {
+      router.replace('/' + lang + '/onboarding');
+    }
+  }, [lang, router]);
+
+  if (!isMounted) {
+    return null;
+  }
   return (
     <Box
       width={'100%'}
       height={'100%'}
-      style={{ display: 'flex', justifyContent: 'center' }}
+      style={{
+        display: 'flex',
+        justifyContent: 'center',
+      }}
     >
-      <Dashboard lang={lang} />
+      <IntroScreen dict={dict} router={router} />
     </Box>
   );
 };
